move trips-layer from sample-layers into trips app
