--- examples/sample-layers/index.js
@@ -1,3 +1,0 @@
-export {default as EnhancedChoroplethLayer} from './enhanced-choropleth-layer';
-export {default as PlotLayer} from './plot-layer';
-export {default as TripsLayer} from './trips-layer';
