+++ examples/trips/trips-layer/trips-layer-fragment.glsl.js
