+++ examples/trips/trips-layer/trips-layer-vertex.glsl.js
